updates to project to make it compile for Mac as well as iOS

+++ src/Frameworks/PGClientKit/NSURL+PGAdditions.m
@@ -75,7 +75,7 @@
 	if(port==0) {
 		return @"";
 	} else {
-		return [NSString stringWithFormat:@":%ld",port];
+		return [NSString stringWithFormat:@":%ld",(unsigned long)port];
 	}
 }
